Добавил JS-модули
